ensure trampolines use PanicTrap
